+++ R/1_model_parameters.R
@@ -27,8 +27,10 @@
 #'
 #' @param model Statistical Model.
 #' @param ... Arguments passed to or from other methods. Non-documented
-#'   arguments are \code{digits}, \code{p_digits} and \code{ci_digits} to set
-#'   the number of digits for the output. See 'Examples' in
+#'   arguments are \code{digits}, \code{p_digits}, \code{ci_digits} and
+#'   \code{footer_digits} to set the number of digits for the output.
+#'   \code{group} can also be passed to the \code{print()} method. See details
+#'   in \code{\link{print.parameters_model}} and 'Examples' in
 #'   \code{\link{model_parameters.default}}.
 #'
 #' @seealso \code{\link[insight:standardize_names]{standardize_names()}} to